% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Additional_functions.R
\name{ar1_cor}
\alias{ar1_cor}
\title{AR(1) correlation matrix}
\usage{
ar1_cor(n, rho)
}
\arguments{
\item{n}{Integer \eqn{\ge 1}. Matrix dimension.}

\item{rho}{Numeric in \eqn{(-1, 1)}. AR(1) correlation parameter.}
}
\value{
A numeric \eqn{n \times n} correlation matrix with entries \eqn{\rho^{|i-j|}}.
}
\description{
Construct an \eqn{n \times n} AR(1) correlation matrix with parameter \code{rho}.
}
