% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Additional_functions.R
\name{generate_group_ind}
\alias{generate_group_ind}
\title{Create a grouping index}
\usage{
generate_group_ind(p, G)
}
\arguments{
\item{p}{Total number of coefficients including intercept.}

\item{G}{Number of groups.}
}
\value{
Integer vector of group indices of length \code{p - 1}.
}
\description{
Split \code{p - 1} non-intercept predictors into \code{G} groups, as evenly as possible.
If division is uneven, earlier groups receive one extra predictor.
}
