% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Additional_functions.R
\name{generate_group_ind_singletons}
\alias{generate_group_ind_singletons}
\title{Assign feature groups for structured variable selection  with singleton consideration.}
\usage{
generate_group_ind_singletons(p, G, singletons = TRUE)
}
\arguments{
\item{p}{Integer. Total number of features (including intercept).}

\item{G}{Number of groups (assumes equal group sizes).}

\item{singletons}{Logical. If TRUE, (p/2)-1 remaining features are assigned to singleton groups.}
}
\value{
A list containing:
\item{group_id}{Integer vector of length \code{p - 1} mapping each feature to its group.}
}
\description{
This function assigns features to groups based on a specified vector of group sizes (\code{Mg}).
If \code{singletons = TRUE}, any remaining features not covered by \code{Mg} are assigned to singleton groups.
}
\examples{
generate_group_ind_singletons(p = 200, G = 20, singletons = TRUE)
}
