% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassNB_main.R
\name{grassNB}
\alias{grassNB}
\title{Main parent function for fitting different models}
\usage{
grassNB(
  X,
  y,
  group_ind = NULL,
  NeighborhoodList = NULL,
  which.prior = c("HS", "SS"),
  niter = 10000,
  verbose = "TRUE",
  pop_col = NULL,
  scale = NULL,
  pi_star = 0.2,
  alpha0 = 10
)
}
\arguments{
\item{X}{is the matrix of features}

\item{y}{is the vector of the count outcome}

\item{group_ind}{is the vector of feature grouping indicators}

\item{NeighborhoodList}{is an object of class \code{nb}, typically created with \code{\link[spdep]{poly2nb}}. Each element of the list corresponds to a spatial unit and contains the indices of its neighboring units, as defined by the chosen contiguity rile (e.g queen or rook). Used to encode the spatial adjacency structure for analysis.}

\item{which.prior}{denotes the model to be used, "HS" or "SS", horseshoe and spike-and-slab respectively}

\item{niter}{is the number of MCMC iterations}

\item{verbose}{is TRUE or FALSE, to show or suppress progression updates}

\item{pop_col}{is a vector of the population of each spatial unit for the model offset, if modeling with space}

\item{scale}{is an integer of the desired scale for outcome rates (e.g 100000 for per 100,000 rates)}

\item{pi_star}{is the prior probability of there being any active feature in a group (when grouping features and using spike and slab)}

\item{alpha0}{is the concentration parameter for the Beta hyper parameter for variable inclusion probability (when grouping features and using spike and slab). Lower values allow for more uncertainty and flexibility.}
}
\value{
It returns the estimated beta's, r, shrinkage parameters, phi, Delta and pDelta (if applicable )
}
\description{
Main parent function for fitting different models
}
