% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Additional_functions.R
\name{phi_True_func}
\alias{phi_True_func}
\title{Sample spatial random effects}
\usage{
phi_True_func(nu_vals, Q, n_space)
}
\arguments{
\item{nu_vals}{Vector of precision parameters.}

\item{Q}{Precision matrix (e.g., ICAR).}

\item{n_space}{Dimension of the spatial domain.}
}
\value{
A list of centered \code{phi} vectors, one per \code{nu}.
}
\description{
Draw spatial random effects \eqn{\phi} from a canonical MVN with precision
\eqn{\Lambda = \nu (Q + \epsilon I)} for each value of \code{nu}. Enforces a
sum-to-zero constraint.
}
