% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NB_MCMC_group_update.R
\name{VS_Group}
\alias{VS_Group}
\title{Function for the NB variable selection (HS or SS) with group updates (with or without space)}
\usage{
VS_Group(
  K,
  y,
  group_ind,
  NeighborhoodList,
  which.prior,
  niter,
  verbose,
  pop_col,
  scale
)
}
\arguments{
\item{K}{Design matrix including an intercept column (first column all 1s)#'}

\item{y}{is the vector of the count outcome}

\item{group_ind}{is the vector of feature grouping indicators}

\item{NeighborhoodList}{is an object of class \code{nb}, typically created with \code{\link[spdep]{poly2nb}}. Each element of the list corresponds to a spatial unit and contains the indices of its neighboring units, as defined by the chosen contiguity rile (e.g queen or rook). Used to encode the spatial adjacency structure for analysis.}

\item{which.prior}{denotes the model to be used, "HS" or "SS", horseshoe and spike-and-slab respectively}

\item{niter}{is the number of MCMC iterations}

\item{verbose}{is TRUE or FALSE, to show or suppress progression updates}

\item{pop_col}{is a vector of the population of each spatial unit for the model offset, if modeling with space}

\item{scale}{is an integer of the desired scale for outcome rates (e.g 100000 for per 100,000 rates)}
}
\value{
It returns the estimated beta's, r, shrinkage parameters, phi, Delta and pDelta (if applicable )
}
\description{
Function for the NB variable selection (HS or SS) with group updates (with or without space)
}
